app/requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
crispy-bootstrap5==2023.10
Django==5.0
django-crispy-forms==2.1
exceptiongroup==1.2.0
gunicorn==21.2.0
iniconfig==2.0.0
packaging==23.2
pluggy==1.3.0
psycopg==3.1.14
pyaml==23.9.7
pydantic==2.5.2
pydantic_core==2.14.5
pytest==7.4.3
pytest-django==4.7.0
PyYAML==6.0.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
uuid==1.30
